Further refactor and avoid cycles.
